# src/swagger.yaml
openapi: 3.0.3
info:
  title: NetScopeNMS REST API
  description: |
    AI 기반 네트워크 관리 시스템 (Network Management System) API 문서
    
    ## 주요 기능
    - **장비 관리**: 네트워크 장비 CRUD, SNMP 연결 테스트, 인터페이스 검색
    - **메트릭 수집**: 실시간 및 이력 데이터 조회 (CPU, 메모리, 트래픽 등)
    - **알람 관리**: 임계값 기반 알람 생성 및 관리
    - **AI 분석**: OpenAI 기반 근본 원인 분석, 예측, 리포트 생성
    - **사용자 관리**: JWT 기반 인증, 역할 기반 권한 관리
    
    ## 인증
    Bearer Token (JWT) 방식을 사용합니다.
    ```
    Authorization: Bearer <token>
    ```
    
    ## 오류 응답 형식
    ```json
    {
      "success": false,
      "error": "Error type",
      "message": "상세 메시지",
      "details": [] // 선택적 상세 정보
    }
    ```
    
  version: 1.0.0
  contact:
    name: NetScopeNMS Team
    email: support@netscopenms.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.netscopenms.com/api/v1
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /alarms:
    get:
      tags:
        - Alarms
      summary: 알람 목록 조회
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, acknowledged, resolved, all]
            default: all
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, critical]
        - name: deviceId
          in: query
          schema:
            type: integer
        - name: metricType
          in: query
          schema:
            type: string
            enum: [cpu, memory, traffic_in, traffic_out, bandwidth_util, errors_in, errors_out, connectivity]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, severity, status, occurrence_count]
            default: created_at
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
      responses:
        '200':
          description: 알람 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      alarms:
                        type: array
                        items:
                          $ref: '#/components/schemas/Alarm'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
...
  /alarms/bulk-acknowledge:
    patch:
      tags:
        - Alarms
      summary: 알람 일괄 확인 처리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - alarmIds
              properties:
                alarmIds:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
                note:
                  type: string
                  example: 점검 완료
      responses:
        '200':
          description: 처리 완료
  /alarms/bulk-resolve:
    patch:
      tags:
        - Alarms
      summary: 알람 일괄 해결 처리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - alarmIds
              properties:
                alarmIds:
                  type: array
                  items:
                    type: integer
                note:
                  type: string
      responses:
        '200':
          description: 처리 완료
  /alarms/{id}:
    get:
      tags:
        - Alarms
      summary: 알람 상세 조회
      parameters:
        - $ref: '#/components/parameters/AlarmId'
      responses:
        '200':
          description: 알람 상세
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      alarm:
                        $ref: '#/components/schemas/AlarmDetail'
  /alarms/{id}/acknowledge:
    patch:
      tags:
        - Alarms
      summary: 알람 확인 처리
      parameters:
        - $ref: '#/components/parameters/AlarmId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  description: 확인 메모
      responses:
        '200':
          description: 확인 처리 완료
  /alarms/{id}/resolve:
    patch:
      tags:
        - Alarms
      summary: 알람 해결 처리
      parameters:
        - $ref: '#/components/parameters/AlarmId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  description: 해결 메모
      responses:
        '200':
          description: 해결 처리 완료

  # ==================== 알람 규칙 ====================
  /alarms/rules:
    get:
      tags:
        - Alarm Rules
      summary: 알람 규칙 목록 조회
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: isEnabled
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: 알람 규칙 목록
    post:
      tags:
        - Alarm Rules
      summary: 알람 규칙 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlarmRuleCreate'
      responses:
        '201':
          description: 생성 성공
  /alarms/rules/{id}:
    get:
      tags:
        - Alarm Rules
      summary: 알람 규칙 상세 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 규칙 상세
    put:
      tags:
        - Alarm Rules
      summary: 알람 규칙 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlarmRuleUpdate'
      responses:
        '200':
          description: 수정 성공
    delete:
      tags:
        - Alarm Rules
      summary: 알람 규칙 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 삭제 성공
  /alarms/rules/{id}/toggle:
    patch:
      tags:
        - Alarm Rules
      summary: 알람 규칙 활성화/비활성화 토글
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 토글 성공

  # ==================== AI 분석 ====================
  /ai/statistics:
    get:
      tags:
        - AI Analysis
      summary: AI 분석 통계
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [24h, 7d, 30d]
            default: 7d
      responses:
        '200':
          description: 분석 통계
  /ai/history:
    get:
      tags:
        - AI Analysis
      summary: AI 분석 이력 조회
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: deviceId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [alarm_rca, prediction, daily_report, weekly_report, anomaly_detection]
      responses:
        '200':
          description: 분석 이력 목록
  /ai/history/{id}:
    get:
      tags:
        - AI Analysis
      summary: AI 분석 상세 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 분석 결과 상세
    delete:
      tags:
        - AI Analysis
      summary: AI 분석 결과 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 삭제 성공
  /ai/devices/{deviceId}/analyze:
    post:
      tags:
        - AI Analysis
      summary: 장비 AI 분석 요청
      parameters:
        - $ref: '#/components/parameters/DeviceIdPath'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                forceRefresh:
                  type: boolean
                  default: false
                  description: 캐시 무시 여부
      responses:
        '200':
          description: 분석 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIAnalysisResponse'
  /ai/devices/{deviceId}/predict:
    post:
      tags:
        - AI Analysis
      summary: 장비 예측 분석
      parameters:
        - $ref: '#/components/parameters/DeviceIdPath'
      responses:
        '200':
          description: 예측 결과
  /ai/alarms/{alarmId}/analyze:
    post:
      tags:
        - AI Analysis
      summary: 알람 근본 원인 분석
      parameters:
        - name: alarmId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 분석 결과
...
  /ai/cache/clear:
    post:
      tags:
        - AI Analysis
      summary: AI 분석 캐시 초기화 (관리자)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId:
                  type: integer
                  description: 특정 장비만 초기화
                type:
                  type: string
                  description: 특정 분석 타입만 초기화
      responses:
        '200':
          description: 캐시 초기화 완료
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT 토큰을 사용한 인증
        
        로그인 후 받은 토큰을 `Authorization: Bearer <token>` 형식으로 전송
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 페이지 번호
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 페이지당 항목 수
...
    DeviceIdPath:
      name: deviceId
      in: path
      required: true
      schema:
        type: integer
      description: 장비 ID
...
    AlarmId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: 알람 ID
  schemas:
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 전체 항목 수
        page:
          type: integer
          description: 현재 페이지
        limit:
          type: integer
          description: 페이지당 항목 수
        totalPages:
          type: integer
          description: 전체 페이지 수
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user, viewer]
        isActive:
          type: boolean
        lastLogin:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Alarm:
      type: object
      properties:
        id:
          type: integer
        device_id:
          type: integer
        interface_id:
          type: integer
        rule_id:
          type: integer
        severity:
          type: string
          enum: [info, warning, critical]
        status:
          type: string
          enum: [active, acknowledged, resolved]
        title:
          type: string
        message:
          type: string
        metric_type:
          type: string
        current_value:
          type: number
        threshold_value:
          type: number
        occurrence_count:
          type: integer
        created_at:
          type: string
          format: date-time
        device:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            ip_address:
              type: string
    AlarmDetail:
      allOf:
        - $ref: '#/components/schemas/Alarm'
        - type: object
          properties:
            acknowledged_at:
              type: string
              format: date-time
            acknowledged_by:
              type: integer
            acknowledgement_note:
              type: string
            resolved_at:
              type: string
              format: date-time
            resolved_by:
              type: integer
            resolution_note:
              type: string
            acknowledgedByUser:
              $ref: '#/components/schemas/User'
            resolvedByUser:
              $ref: '#/components/schemas/User'
    AlarmRuleCreate:
      type: object
      required:
        - name
        - metric_type
        - threshold_critical
      properties:
        name:
          type: string
          maxLength: 100
          example: CPU 높은 사용률 경고
        description:
          type: string
          example: CPU 사용률이 임계값을 초과하면 알람 발생
        metric_type:
          type: string
          enum: [cpu, memory, traffic_in, traffic_out, bandwidth_util, errors_in, errors_out]
          example: cpu
        condition_operator:
          type: string
          enum: [gt, gte, lt, lte, eq, neq]
          default: gt
          description: "gt: >, gte: >=, lt: <, lte: <=, eq: =, neq: !="
        threshold_warning:
          type: number
          example: 70
          description: 경고 임계값
        threshold_critical:
          type: number
          example: 90
          description: 긴급 임계값
        duration_seconds:
          type: integer
          default: 0
          description: 조건 지속 시간 (초)
        cooldown_seconds:
          type: integer
          default: 300
          description: 알람 재발생 방지 시간 (초)
        device_filter:
          type: object
          description: 적용 대상 장비 필터
    AlarmRuleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        threshold_warning:
          type: number
        threshold_critical:
          type: number
        duration_seconds:
          type: integer
        cooldown_seconds:
          type: integer
        is_enabled:
          type: boolean
    AIAnalysisResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            deviceId:
              type: integer
            deviceName:
              type: string
            cached:
              type: boolean
              description: 캐시된 결과 여부
            result:
              type: object
              description: AI 분석 결과 (JSON 형식)
            analysisId:
              type: integer
            tokensUsed:
              type: integer
...